new/fix: reworked the whole XFCE setup
xfconf-query needs TIME! damn.. that was a hard finding because when you set
something by xfconf-query it will not get written immediately it need an unknown
amount of time to do so.
Now we check the md5sum before actually setting something (if the file is not there
we set the md5sum to 0) and continue only when the file gets modified/created.
F_DOMD5 will create the md5
F_FILEWAIT will compare the given md5 and waits until it changes