new: introducing the TEST area
due to an issue currently occurred I thought we should have a test section
where we ensure that all files are in place we require and everything else
which can be checked, of course.
Currently we check for file existence only but that can be extended as we
like. important thing is that we need to make USE of it by adding / extending
that area whenever possible