fix: randomly wrong calc for ISO size
before we used 'du' to calculate the size of the ISO -
the same way archiso does btw. the problem: especially on
full clean builds the ISO size get reported as 1 GB by 'du'
even when less then 700 MB only.
This was no problem before persistent mode but as I use the
ISO file size to calculate the persistent storage it resutls in
totally wrong ISOs at the end.
stat is bulletproof here so I switched to it instead.

Change-Id: I21b983fcd6356a00c75c6e8ea65bf00f24e7d90a