new: disk space check
before installing a package with one of the installers
the GUI wrapper will check for available disk space plus
the remote reported installed package size.
Will abort when the local disk space is not enough.
when a pkg does not provide its size (e.g. Teamviewer doesn't)
it will warn the user with a hint for the available disk size
and will give an option to continue regardless

Change-Id: I084472978092a2ed1f74a03ef5685ee027df53c3